lab_6_restful/requirements.txt
alembic==1.15.1
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
fastapi==0.115.11
flasgger==0.9.7.1
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonify==0.5
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mistune==3.1.3
motor==3.7.0
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-mongo==3.0.0
pydantic_core==2.27.2
pymongo==4.12.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.0
typing_extensions==4.12.2
uv==0.6.5
uvicorn==0.34.0
Werkzeug==3.1.3
